fix: update confusion matrix and box spreading to ensure dtype consistency
- metrics/detection.py: adjust confusion matrix to use `dtype=int32`, updated doctests accordingly.
- annotators/core.py: explicitly cast to `npt.NDArray[np.float32]` for box spreading operation.

+++ src/supervision/metrics/detection.py
@@ -143,8 +143,8 @@ class ConfusionMatrix:
             ...     classes=['person']
             ... )
             >>> confusion_matrix.matrix
-            array([[1., 0.],
-                   [0., 0.]])
+            array([[1, 0],
+                   [0, 0]], dtype=int32)
 
             ```
         """
@@ -217,9 +217,9 @@ class ConfusionMatrix:
             ...     classes=['person', 'dog']
             ... )
             >>> confusion_matrix.matrix
-            array([[1., 0., 1.],
-                   [0., 1., 0.],
-                   [1., 0., 0.]])
+            array([[1, 0, 1],
+                   [0, 1, 0],
+                   [1, 0, 0]], dtype=int32)
 
             ```
         """
